automations/example/requirements.txt
psycopg2-binary==2.9.9
pika==1.3.2
requests==2.31.0
python-dotenv==1.0.0
